pyproject.toml
[tool.poetry]
name = "aws-strava-stats"
version = "0.1.0"
description = ""
authors = ["David Enthoven <david.enthoven@live.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
constructs = "^10.3.0"
aws-cdk-lib = "^2.147.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
boto3 = "^1.34.136"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
